+++ gemstone/src/payment/solana_pay.rs
@@ -15,7 +15,7 @@ pub struct SolanaPayLabel {
     pub label: String,
     #[serde(rename = "icon")]
     pub icon_url: String,
-    pub error: Option<String>,
+    pub title: Option<String>,
 }
 
 #[derive(Debug, Deserialize, uniffi::Record)]
@@ -38,9 +38,6 @@ impl SolanaPay {
         let label = serde_json::from_slice::<SolanaPayLabel>(&response).map_err(|_| GemstoneError::AnyError {
             msg: "Failed to get solana pay label and icon".into(),
         })?;
-        if let Some(error) = label.error {
-            return Err(GemstoneError::AnyError { msg: error });
-        }
         Ok(label)
     }
