+++ gemstone/tests/ios/GemTest/GemTest/GemTestApp.swift
@@ -1,22 +1,16 @@
-//
-//  GemTestApp.swift
-//  GemTest
-//
-//  Created by magician on 20/12/23.
-//
+// Copyright (c). Gem Wallet. All rights reserved.
 
-import SwiftUI
 import Gemstone
+import SwiftUI
 
 class AppDelegate: NSObject, UIApplicationDelegate {
-    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil) -> Bool {
         return true
     }
 }
 
 @main
 struct GemTestApp: App {
-    
     @UIApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
 
     var body: some Scene {
